syntax

+++ commands/supprime.js
@@ -1,11 +1,11 @@
-const Discord = require("discord.js");
+// eslint-disable-next-line no-unused-vars
+const Discord = require('discord.js')
 
 module.exports.run = async (bot, message, args) => {
   message.guild.roles.forEach(role => {
-    if (role.name == message.author.username)
-      role.delete();
-  });
-  return message.reply("bon normalement ta/tes couleurs sont supprimées mdr.");
+    if (role.name === message.author.username) { role.delete() }
+  })
+  return message.reply('bon normalement ta/tes couleurs sont supprimées mdr.')
 }
 
 module.exports.help = {
